Moved the SearchError to utility and removed the module structure
