+++ requirements.txt
@@ -5,3 +5,4 @@ requests==2.24.0
 termcolor==1.1.0
 urllib3==1.25.10
 loldb==0.1.5
+rich
