--- src/arguments/__init__.py
@@ -1,11 +1,0 @@
-""" 
-                    GNU GENERAL PUBLIC LICENSE
-                       Version 3, 29 June 2007
-
- Copyright (C) 2007 Free Software Foundation, Inc. <https://fsf.org/>
- Everyone is permitted to copy and distribute verbatim copies
- of this license document, but changing it is not allowed.
-"""
-
-from .search import SearchError, Search
-from .utility import Utility
